+++ rubato/utils/rendering/font.py
@@ -78,9 +78,9 @@ class Font:
         self._color = new
         self._font = sdl2.ext.FontTTF(self._font_path, self._size, self._color.to_tuple())
 
-    def generate_surface(self, text: str, align: str, width: int | float = 0) -> sdl2.SDL_Surface:
+    def _generate(self, text: str, align: str, width: int | float = 0) -> sdl2.SDL_Surface:
         """
-        Generate the surface containing the text.
+        Generate a surface containing the text.
 
         Args:
             text: The text to render.