+++ tests/init_test.py
@@ -25,7 +25,7 @@ def test_begin_init(monkeypatch, rub):
     # pylint: disable=unused-argument
     assert rubato.Game.state == rubato.Game.STOPPED
     loop = Mock()
-    monkeypatch.setattr(rubato.Game, "start", loop)
+    monkeypatch.setattr(rubato.Game, "_start", loop)
     rubato.Display.hidden = False
     rubato.begin()
     loop.assert_called_once()  # This code is reachable because of the monkeypatch